L1 Virtual interrupt handler with interface
And also Move away from nested namespace inclusion

Resolves #629
Resolves #633
