--- library/L0_Platform/example/example_memory_map_c.h
@@ -1,3 +1,0 @@
-#pragma once
-// Replace this file with the memory map definitions file for your
-// microcontroller.
