--- library/L0_Platform/example/interrupt.cpp
@@ -1,3 +1,0 @@
-#include "L0_Platform/example/interrupt.hpp"
-
-// This depends greatly on the platform you are working with.
