--- library/L0_Platform/example/interrupt.hpp
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "L0_Platform/interrupt.hpp"
-
-// Incorporate any platform specific interrupt code that should be available
-// in the startup.cpp file.
