+++ library/L0_Platform/lpc17xx/startup.cpp
@@ -35,11 +35,11 @@
 #include <cstring>
 #include <iterator>
 
-#include "L0_Platform/lpc17xx/interrupt.hpp"
 #include "L0_Platform/ram.hpp"
 #include "L0_Platform/startup.hpp"
 #include "L1_Peripheral/cortex/dwt_counter.hpp"
 #include "L1_Peripheral/cortex/system_timer.hpp"
+#include "L1_Peripheral/interrupt.hpp"
 #include "L1_Peripheral/lpc17xx/system_controller.hpp"
 #include "L1_Peripheral/lpc17xx/timer.hpp"
 #include "L1_Peripheral/lpc17xx/uart.hpp"
@@ -50,6 +50,7 @@
 // Private namespace to make sure that these do not conflict with other globals
 namespace
 {
+using sjsu::cortex::InterruptController;
 // Create LPC40xx system controller to be used by low level initialization.
 sjsu::lpc17xx::SystemController system_controller;
 // Create timer0 to be used by lower level initialization for uptime calculation
@@ -59,6 +60,8 @@ sjsu::lpc17xx::Uart uart0(sjsu::lpc17xx::UartPort::kUart0, system_controller);
 // System timer is used to count milliseconds of time and to run the RTOS
 // scheduler.
 sjsu::cortex::SystemTimer system_timer(system_controller);
+// Cortex NVIC interrupt controller used to setup FreeRTOS ISRs
+sjsu::cortex::InterruptController interrupt_controller;
 
 int Lpc17xxStdOut(const char * data, size_t length)
 {
@@ -85,12 +88,35 @@ extern "C"
   // The entry point for the application.
   // main() is the entry point for newlib based applications
   extern int main();
+
+  // External declaration for the pointer to the stack top from the linker
+  // script
+  extern void StackTop(void);
+  // These are defined after the compilation of the FreeRTOS port for Cortex M4F
+  // These will link to those definitions.
+  extern void xPortPendSVHandler(void);   // NOLINT
+  extern void vPortSVCHandler(void);      // NOLINT
+  extern void xPortSysTickHandler(void);  // NOLINT
+  // Forward declaration of the default handlers. These are aliased.
+  // When the application defines a handler (with the same name), the
+  // application's handler will automatically take precedence over these weak
+  // definitions.
+  extern void HardFaultHandler(void);
+
   void vPortSetupTimerInterrupt(void)  // NOLINT
   {
+    interrupt_controller.Register({
+        .interrupt_request_number  = sjsu::cortex::SVCall_IRQn,
+        .interrupt_service_routine = vPortSVCHandler,
+    });
+    interrupt_controller.Register({
+        .interrupt_request_number  = sjsu::cortex::PendSV_IRQn,
+        .interrupt_service_routine = xPortPendSVHandler,
+    });
     // Set the SystemTick frequency to the RTOS tick frequency
-    // It is critical that this happens before you set the system_clock, since
-    // The system_timer keeps the time that the system_clock uses to delay
-    // itself.
+    // It is critical that this happens before you set the system_clock,
+    // since The system_timer keeps the time that the system_clock uses to
+    // delay itself.
     system_timer.SetInterrupt(xPortSysTickHandler);
   }
 }
@@ -128,9 +154,9 @@ extern "C" void ResetIsr()
   // bootloader function launches this ISR manually, but it never returns thus
   // it never cleans up the memory it uses. To get that memory back, we have
   // to manually move the stack pointers back to the top of stack.
-  const intptr_t kTopOfStack = reinterpret_cast<intptr_t>(&StackTop);
-  sjsu::lpc17xx::__set_PSP(kTopOfStack);
-  sjsu::lpc17xx::__set_MSP(kTopOfStack);
+  const uint32_t kTopOfStack = reinterpret_cast<intptr_t>(&StackTop);
+  sjsu::cortex::__set_PSP(kTopOfStack);
+  sjsu::cortex::__set_MSP(kTopOfStack);
 
   sjsu::SystemInitialize();
   InitializePlatform();
@@ -155,3 +181,65 @@ extern "C" void ResetIsr()
   }
   sjsu::Halt();
 }
+
+// The Interrupt vector table.
+// This relies on the linker script to place at correct location in memory.
+SJ2_SECTION(".isr_vector")
+// NOLINTNEXTLINE(readability-identifier-naming)
+const sjsu::IsrPointer kInterruptVectorTable[] = {
+  // Core Level - CM4
+  &StackTop,                           // 0, The initial stack pointer
+  ResetIsr,                            // 1, The reset handler
+  InterruptController::LookupHandler,  // 2, The NMI handler
+  HardFaultHandler,                    // 3, The hard fault handler
+  InterruptController::LookupHandler,  // 4, The MPU fault handler
+  InterruptController::LookupHandler,  // 5, The bus fault handler
+  InterruptController::LookupHandler,  // 6, The usage fault handler
+  nullptr,                             // 7, Reserved
+  nullptr,                             // 8, Reserved
+  nullptr,                             // 9, Reserved
+  nullptr,                             // 10, Reserved
+  InterruptController::LookupHandler,  // 11, SVCall handler
+  InterruptController::LookupHandler,  // 12, Debug monitor handler
+  nullptr,                             // 13, Reserved
+  InterruptController::LookupHandler,  // 14, FreeRTOS PendSV Handler
+  InterruptController::LookupHandler,  // 15, The SysTick handler
+  // Chip Level - LPC17xx
+  InterruptController::LookupHandler,  // 16, 0x40 - WDT
+  InterruptController::LookupHandler,  // 17, 0x44 - TIMER0
+  InterruptController::LookupHandler,  // 18, 0x48 - TIMER1
+  InterruptController::LookupHandler,  // 19, 0x4c - TIMER2
+  InterruptController::LookupHandler,  // 20, 0x50 - TIMER3
+  InterruptController::LookupHandler,  // 21, 0x54 - UART0
+  InterruptController::LookupHandler,  // 22, 0x58 - UART1
+  InterruptController::LookupHandler,  // 23, 0x5c - UART2
+  InterruptController::LookupHandler,  // 24, 0x60 - UART3
+  InterruptController::LookupHandler,  // 25, 0x64 - PWM1
+  InterruptController::LookupHandler,  // 26, 0x68 - I2C0
+  InterruptController::LookupHandler,  // 27, 0x6c - I2C1
+  InterruptController::LookupHandler,  // 28, 0x70 - I2C2
+  InterruptController::LookupHandler,  // 29, 0x74 - SPI0
+  InterruptController::LookupHandler,  // 30, 0x78 - SSP0
+  InterruptController::LookupHandler,  // 31, 0x7c - SSP1
+  InterruptController::LookupHandler,  // 32, 0x80 - PLL0 (Main PLL)
+  InterruptController::LookupHandler,  // 33, 0x84 - RTC
+  InterruptController::LookupHandler,  // 34, 0x88 - EINT0
+  InterruptController::LookupHandler,  // 35, 0x8c - EINT1
+  InterruptController::LookupHandler,  // 36, 0x90 - EINT2
+  InterruptController::LookupHandler,  // 37, 0x94 - EINT3
+  InterruptController::LookupHandler,  // 38, 0x98 - ADC
+  InterruptController::LookupHandler,  // 39, 0x9c - BOD
+  InterruptController::LookupHandler,  // 40, 0xA0 - USB
+  InterruptController::LookupHandler,  // 41, 0xa4 - CAN
+  InterruptController::LookupHandler,  // 42, 0xa8 - GP DMA
+  InterruptController::LookupHandler,  // 43, 0xac - I2S
+  InterruptController::LookupHandler,  // 44, 0xb0 - Ethernet
+  InterruptController::LookupHandler,  // 45, 0xb4 - RIT
+  InterruptController::LookupHandler,  // 46, 0xb8 - Motor Control PWM
+  InterruptController::LookupHandler,  // 47, 0xbc - Quadrature Encoder
+  InterruptController::LookupHandler,  // 48, 0xc0 - PLL1 (USB PLL)
+  InterruptController::LookupHandler,  // 49, 0xc4 - USB Activity interrupt to
+                                       // wakeup
+  InterruptController::LookupHandler,  // 50, 0xc8 - CAN Activity interrupt to
+                                       // wakeup
+};
