+++ library/L1_Peripheral/cortex/L1_cortex.mk
@@ -1,2 +1,3 @@
 TESTS += $(LIBRARY_DIR)/L1_Peripheral/cortex/test/dwt_counter_test.cpp
 TESTS += $(LIBRARY_DIR)/L1_Peripheral/cortex/test/system_timer_test.cpp
+# TESTS += $(LIBRARY_DIR)/L1_Peripheral/cortex/interrupt_test.cpp
