+++ library/L1_Peripheral/lpc40xx/L1_lpc40xx.mk
@@ -7,3 +7,4 @@ TESTS += $(LIBRARY_DIR)/L1_Peripheral/lpc40xx/test/pin_test.cpp
 TESTS += $(LIBRARY_DIR)/L1_Peripheral/lpc40xx/test/pwm_test.cpp
 TESTS += $(LIBRARY_DIR)/L1_Peripheral/lpc40xx/test/spi_test.cpp
 TESTS += $(LIBRARY_DIR)/L1_Peripheral/lpc40xx/test/uart_test.cpp
+TESTS += $(LIBRARY_DIR)/L1_Peripheral/lpc40xx/test/system_controller_test.cpp
