+++ library/L1_Peripheral/lpc40xx/can.hpp
@@ -4,10 +4,11 @@
 #include <queue.h>
 
 #include <initializer_list>
+#include <cstring>
 
 #include "L1_Peripheral/can.hpp"
 
-#include "L0_Platform/lpc40xx/interrupt.hpp"
+#include "L1_Peripheral/cortex/interrupt.hpp"
 #include "L0_Platform/lpc40xx/LPC40xx.h"
 #include "L1_Peripheral/lpc40xx/pin.hpp"
 #include "L1_Peripheral/lpc40xx/system_controller.hpp"
@@ -282,7 +283,78 @@ class Can final : public sjsu::Can
     kAcceptAllMessages          = 0x02,
   };
 
+  static void ProcessIrq()
+  {
+    TxMessage_t * message_ptr;
+    bool transmit_interrupt, receive_interrupt;
+
+    for (uint8_t controller = 0; controller < kNumberOfControllers;
+         controller++)
+    {
+      if (is_controller_initialized[controller] == true)
+      {
+        // Capture interrupt register
+        // Bits are cleared when you read the ICR register
+        interrupts[controller].ICR = can_registers[controller]->ICR;
+
+        transmit_interrupt = interrupts[controller].flags.tx_buffer_1_ready ||
+                             interrupts[controller].flags.tx_buffer_2_ready ||
+                             interrupts[controller].flags.tx_buffer_3_ready;
+
+        receive_interrupt = interrupts[controller].flags.rx_buffer_full;
+
+        if (transmit_interrupt == true)
+        {
+          if (xQueueIsQueueEmptyFromISR(transmit_queue[controller]) == pdFALSE)
+          {
+            // dequeue message pointer
+            xQueueReceiveFromISR(transmit_queue[controller], &message_ptr, 0);
+            // send it to the available buffer
+            if (interrupts[controller].flags.tx_buffer_1_ready)
+            {
+              WriteMessageToBuffer(
+                  message_ptr, kBuffer1, static_cast<Controllers>(controller));
+            }
+            else if (interrupts[controller].flags.tx_buffer_2_ready)
+            {
+              WriteMessageToBuffer(
+                  message_ptr, kBuffer2, static_cast<Controllers>(controller));
+            }
+            else if (interrupts[controller].flags.tx_buffer_3_ready)
+            {
+              WriteMessageToBuffer(
+                  message_ptr, kBuffer3, static_cast<Controllers>(controller));
+            }
+            else
+            {
+              // will never occur
+            }
+          }
+        }
+
+        if (receive_interrupt == true)
+        {
+          ReadMessageFromBuffer(static_cast<Controllers>(controller));
+        }
+      }
+    }
+  }
+
+  static void CanIrqHandler()
+  {
+    ProcessIrq();
+  }
+
+  inline static const InterruptController::RegistrationInfo_t
+      kCanInterruptInfo = {
+        .interrupt_request_number  = CAN_IRQn,
+        .interrupt_service_routine = &CanIrqHandler,
+        .enable_interrupt          = true,
+        .priority                  = 5,
+      };
+
   static const sjsu::lpc40xx::SystemController kDefaultSystemController;
+  static const sjsu::cortex::InterruptController kCortexInterruptController;
 
   // Default constructor that defaults to CAN 1
   constexpr Can()
@@ -292,24 +364,27 @@ class Can final : public sjsu::Can
         td_(&td_pin_),
         rd_pin_(Pin(kCan1Port, kRd1PinNumber)),
         td_pin_(Pin(kCan1Port, kTd1PinNumber)),
-        system_controller_(kDefaultSystemController)
+        system_controller_(kDefaultSystemController),
+        interrupt_controller_(kCortexInterruptController)
   {
   }
 
   constexpr Can(
-              Controllers controller,
-              BaudRates baud_rate,
-              sjsu::Pin * td_pin,
-              sjsu::Pin * rd_pin,
-              const sjsu::lpc40xx::SystemController & system_controller =
-                  kDefaultSystemController)
+      Controllers controller,
+      BaudRates baud_rate,
+      sjsu::Pin * td_pin,
+      sjsu::Pin * rd_pin,
+      const sjsu::lpc40xx::SystemController & system_controller =
+          kDefaultSystemController,
+      const sjsu::cortex::InterruptController & = kCortexInterruptController)
       : controller_(controller),
         baud_rate_(baud_rate),
         rd_(td_pin),
         td_(rd_pin),
         rd_pin_(Pin::CreateInactivePin()),
         td_pin_(Pin::CreateInactivePin()),
-        system_controller_(system_controller)
+        system_controller_(system_controller),
+        interrupt_controller_(kCortexInterruptController)
   {
   }
 
@@ -338,7 +413,8 @@ class Can final : public sjsu::Can
       // TODO(#343): Allow the user to configure their own filter.
       EnableAcceptanceFilter();
       EnableInterrupts();
-      RegisterIsr(CAN_IRQn, &CanIrqHandler, true, 5);
+
+      interrupt_controller_.Register(kCanInterruptInfo);
       // Disable reset mode and enter operating mode.
       SetControllerMode(kReset, false);
       is_controller_initialized[controller_] = true;
@@ -356,8 +432,10 @@ class Can final : public sjsu::Can
   // to send a 1 time message. However, if the user is sending cyclic messages,
   // it is assumed that global message handles exist somewhere in the user's app
   // layer since they need to be updated periodically.
-  bool Send(TxMessage_t * const kMessage, uint32_t id,
-            const uint8_t * const kPayload, size_t length) const override
+  bool Send(TxMessage_t * const kMessage,
+            uint32_t id,
+            const uint8_t * const kPayload,
+            size_t length) const override
   {
     bool success = false;
 
@@ -370,7 +448,7 @@ class Can final : public sjsu::Can
     // the ISR (to eventually write the data at this memory location into the
     // transmit buffer), we disable the CAN interrupt to ensure that complete
     // and valid data is written before it is accessed by the ISR.
-    DeregisterIsr(CAN_IRQn);
+    interrupt_controller_.Deregister(CAN_IRQn);
     kMessage->id                           = id;
     kMessage->frame_data.data_length       = (length & 0xF);
     kMessage->frame_data.remote_tx_request = 0;
@@ -378,7 +456,8 @@ class Can final : public sjsu::Can
     {
       kMessage->data.bytes[i] = kPayload[i];
     }
-    RegisterIsr(CAN_IRQn, &CanIrqHandler, true, 5);
+
+    interrupt_controller_.Register(kCanInterruptInfo);
 
     // Check if any buffer is available.
     if (status[controller_].flags.tx_buffer_1_released)
@@ -501,68 +580,6 @@ class Can final : public sjsu::Can
     SetControllerMode(kReset, false);
   }
 
-  static void ProcessIrq()
-  {
-    TxMessage_t * message_ptr;
-    bool transmit_interrupt, receive_interrupt;
-
-    for (uint8_t controller = 0; controller < kNumberOfControllers;
-         controller++)
-    {
-      if (is_controller_initialized[controller] == true)
-      {
-        // Capture interrupt register
-        // Bits are cleared when you read the ICR register
-        interrupts[controller].ICR = can_registers[controller]->ICR;
-
-        transmit_interrupt = interrupts[controller].flags.tx_buffer_1_ready ||
-                             interrupts[controller].flags.tx_buffer_2_ready ||
-                             interrupts[controller].flags.tx_buffer_3_ready;
-
-        receive_interrupt = interrupts[controller].flags.rx_buffer_full;
-
-        if (transmit_interrupt == true)
-        {
-          if (xQueueIsQueueEmptyFromISR(transmit_queue[controller]) == pdFALSE)
-          {
-            // dequeue message pointer
-            xQueueReceiveFromISR(transmit_queue[controller], &message_ptr, 0);
-            // send it to the available buffer
-            if (interrupts[controller].flags.tx_buffer_1_ready)
-            {
-              WriteMessageToBuffer(message_ptr, kBuffer1,
-                                   static_cast<Controllers>(controller));
-            }
-            else if (interrupts[controller].flags.tx_buffer_2_ready)
-            {
-              WriteMessageToBuffer(message_ptr, kBuffer2,
-                                   static_cast<Controllers>(controller));
-            }
-            else if (interrupts[controller].flags.tx_buffer_3_ready)
-            {
-              WriteMessageToBuffer(message_ptr, kBuffer3,
-                                   static_cast<Controllers>(controller));
-            }
-            else
-            {
-              // will never occur
-            }
-          }
-        }
-
-        if (receive_interrupt == true)
-        {
-          ReadMessageFromBuffer(static_cast<Controllers>(controller));
-        }
-      }
-    }
-  }
-
-  static void CanIrqHandler()
-  {
-    ProcessIrq();
-  }
-
   template <typename T>
   void CreateStaticQueues(T static_queue_config)
   {
@@ -580,7 +597,8 @@ class Can final : public sjsu::Can
 
  private:
   [[gnu::always_inline]] static void WriteMessageToBuffer(
-      const TxMessage_t * const kMessage, TxBuffers buffer_number,
+      const TxMessage_t * const kMessage,
+      TxBuffers buffer_number,
       Controllers controller)
   {
     volatile TxMessage_t * ptr_to_buffer_offset;
@@ -614,12 +632,12 @@ class Can final : public sjsu::Can
         break;
     }
 
-    DeregisterIsr(CAN_IRQn);
+    kCortexInterruptController.Deregister(CAN_IRQn);
     ptr_to_buffer_offset->TFI        = kMessage->TFI;
     ptr_to_buffer_offset->id         = kMessage->id;
     ptr_to_buffer_offset->data.qword = kMessage->data.qword;
     can_registers[controller]->CMR   = command;
-    RegisterIsr(CAN_IRQn, &CanIrqHandler, true, 5);
+    kCortexInterruptController.Register(kCanInterruptInfo);
   }
 
   [[gnu::always_inline]] static void ReadMessageFromBuffer(
@@ -632,11 +650,11 @@ class Can final : public sjsu::Can
         reinterpret_cast<volatile RxMessage_t *>(
             &can_registers[controller]->RFS);
 
-    NVIC_DisableIRQ(CAN_IRQn);
+    kCortexInterruptController.Deregister(CAN_IRQn);
     message.RFS        = ptr_to_buffer_offset->RFS;
     message.id         = ptr_to_buffer_offset->id;
     message.data.qword = ptr_to_buffer_offset->data.qword;
-    NVIC_EnableIRQ(CAN_IRQn);
+    kCortexInterruptController.Register(kCanInterruptInfo);
 
     xQueueSendFromISR(receive_queue[controller], &message, 0);
     can_registers[controller]->CMR = kReleaseRxBuffer;
@@ -763,6 +781,7 @@ class Can final : public sjsu::Can
   lpc40xx::Pin rd_pin_;
   lpc40xx::Pin td_pin_;
   const sjsu::lpc40xx::SystemController & system_controller_;
+  const sjsu::cortex::InterruptController & interrupt_controller_;
 };
 }  // namespace lpc40xx
 }  // namespace sjsu
