+++ library/L3_Application/commandline.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <cstdio>
+#include <cstring>
 #include <iterator>
 
 #include "newlib/newlib.hpp"
