+++ library/L3_Application/commands/i2c_command.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <cstdlib>
+#include <cstring>
 #include <iterator>
 #include <tuple>
