+++ library/L3_Application/commands/lpc_system_command.hpp
@@ -3,7 +3,7 @@
 #include <cinttypes>
 #include <iterator>
 
-#include "L0_Platform/lpc40xx/interrupt.hpp"
+#include "L1_Peripheral/cortex/interrupt.hpp"
 #include "L0_Platform/ram.hpp"
 #include "L1_Peripheral/lpc40xx/system_controller.hpp"
 #include "L3_Application/commandline.hpp"
@@ -11,6 +11,7 @@
 
 namespace sjsu
 {
+extern "C" void StackTop(void);
 /// The SystemInfoCommand allows the user to get runtime stats about the
 // processor like, memory left, heap left, etc ...
 class LpcSystemInfoCommand final : public Command
@@ -29,7 +30,7 @@ class LpcSystemInfoCommand final : public Command
     uint32_t peripheral_frequency = system.GetPeripheralFrequency({}) / 1000;
 
     intptr_t top_of_stack         = reinterpret_cast<intptr_t>(&StackTop);
-    intptr_t master_stack_pointer = sjsu::lpc40xx::__get_MSP();
+    intptr_t master_stack_pointer = sjsu::cortex::__get_MSP();
     intptr_t used_stack           = top_of_stack - master_stack_pointer;
     intptr_t remaining_stack      = 0x10000 - used_stack;
     intptr_t head_used            = heap_position - &heap;
